Abstract Interface

This chapter documents the abstract interface to access a result data structure.

Functions	Description

SignalType	Enumeration defining the supported signal types.

rawSignal	Returns signal data given the signal name (required).

names	Return all signal names (required).

timeSignalName	Return the name of the time signal (required).

hasOneTimeSignal	Return true if one time signal present (required).

hasSignal	Inquire whether a signal name is known (optional).

defaultHeading	Return default heading as string (optional).

ModiaResult.SignalType — Type

@enum ModiaResult.SignalType

Defines the type of the signal. Supported values:

ModiaResult.TimeSignal: Time signal (= independent variable).

ModiaResult.Continuous: Piece-wise continuous signal (typically linearly interpolated).

ModiaResult.Clocked: Clocked signal (values are only defined at the corresponding Time signal time instants and have no value in between; the latter might be signaled by piece-wise constant dotted lines).

ModiaResult.rawSignal — Function

(timeSignal, signal, signalType) = ModiaResult.rawSignal(result, name)

Returns

the result time series signal::Vector{AbstractVector}} of name::AbstractString (an element of signal[i][j] is either a Real number (<: Real) or an array of Real numbers (eltype(signal[i][j]) <: Real),

the corresponding timeSignal::Vector{Vector{Real}} of the independent variable,

the information signalType::SignalType that defines how the signal shall be interpolated.

Note, an error shall be raised, if name is not known.

Result signals consist of one or more segments. signal[i][j] is the value of the signal at time instant timeSignal[i][j] in result segment i. timeSignal[i][:] must have monotonically increasing values.

If the signal is a constant with value value, return ([[value, value]], [[t_min, t_max]], timeSignalName, Continuous).

If the signal is the time signal, return (timeSignal, timeSignal, timeSignalName, TimeSignal). The timeSignal might be a dummy vector consisting of the first and last time point in the result (if different timeSignals are present for different signals or if the signal is constant).

signal and timeSignal may have units from package Unitful.

ModiaResult.names — Function

ModiaResult.names(result)

Return a string vector of the signal names that are present in result.

ModiaResult.timeSignalName — Function

ModiaResult.timeSignalName(result)

Return the name of the time signal (default: "time").

ModiaResult.hasOneTimeSignal — Function

ModiaResult.hasOneTimeSignal(result)

Return true if result has one time signal. Return false, if result has two or more time signals.

ModiaResult.hasSignal — Function

ModiaResult.hasSignal(result, name)

Returns true if signal name::AbstractString is available in result.

ModiaResult.defaultHeading — Function

ModiaResult.defaultHeading(result)

Return default heading of result as a string (can be used as default heading for a plot).