Internal

This chapter documents internal functions that are typically only for use of the developers of package ModiaResult.

ModiaResult.getSignalDetailsWithWarning — Function

(signal, timeSignal, timeSignalName, signalType, arrayName, arrayIndices, nScalarSignals) = 
     getSignalWithWarning(result, name)

Call getSignal(result,name) and print a warning message if signal == nothing

ModiaResult.signalLength — Function

signalLength(signal)

Return the total number of values of signal::Vector{AbstractVector}. If signal[i] is nothing or missing, a length of zero is returned.

ModiaResult.hasSameSegments — Function

hasSameSegments(signal1, signal2)

Return true, if the lengths of the segments in signal1 and in signal2 are the same.

ModiaResult.hasDimensionMismatch — Function

hasDimensionMismatch(signal, timeSignal, timeSignalName)

Print a warning message if signalLength(signal) != signalLength(timeSignal) and return true. Otherwise, return false

ModiaResult.prepend! — Function

prepend!(prefix, signalLegend)

Add prefix string in front of every element of the signalLegend string-Vector.