Internal

This chapter documents internal functions that are typically only for use of the developers of package ModiaResult.

ModiaResult.info — Function

info(result)

Return names, units, signal types, values sizes, and element types of the signals that are stored in result as DataFrame table.

ModiaResult.showInfo — Function

showInfo(result)

Print names, units, signal types, values sizes, and element types of the signals that are stored in result as DataFrame table.

ModiaResult.signalInfo — Function

(sigType, nTime, sigSize, sigElType, sigUnit) = signalInfo(result, name)

Return information about a signal, given the name of the signal in result:

sigType::SignalType: Ìnterpolation type of signal.

nTime::Int: Number of signal time points.

sigSize: size(signal[1][1])

sigElType: ustrip( eltype(signal[1][1]) ), that is the element type of the signal without unit.

sigUnit: Unit of signal

If name is defined, but no signal is available (= nothing, missing or zero length), return nTime=0 and nothing for sigSize, sigElType, sigUnit.

ModiaResult.getSignal — Function

(signal, timeSignal, timeSignalName, signalType, arrayName, arrayIndices, nScalarSignals) = getSignal(result, name)

Return the signal defined by name::AbstractString as signal::Vector{Matrix{<:Real}}. name may include an array range, such as "a.b.c[2:3,5]". In this case arrayName is the name without the array indices, such as "a.b.c", arrayIndices is a tuple with the array indices, such as (2:3, 5) and nScalarSignals is the number of scalar signals, such as 3 if arrayIndices = (2:3, 5). Otherwise arrayName = name, arrayIndices=(), nScalarSignals=1.

In case the signal is not known or name cannot be interpreted, (nothing, nothing, nothing, nothing, name, (), 0) is returned.

It is required that the value of the signal at a time instant has either typeof(value) <: Real or typeof(value) = AbstractArray{Real, N}. The following Real types are currently supported:

convert(Float64, eltype(value) is supported (for example Float32, Float64, DoubleFloat, Rational, Int32, Int64, Bool).

Measurements.Measurement{<Type of (1)>}.

MonteCarloMeasurements.StaticParticles{<Type of (1)>}.

MonteCarloMeasurements.Particles{<Type of (1)>}.

ModiaResult.getSignalWithWarning — Function

(signal, timeSignal, timeSignalName, signalType, arrayName, arrayIndices, nScalarSignals) = 
     getSignalWithWarning(result, name)

Call getSignal(result,name) and print a warning message if signal == nothing

ModiaResult.getPlotSignal — Function

(xsig, xsigLegend, ysig, ysigLegend, ysigType) = getPlotSignal(result, xsigName, ysigName)

Given the name of the signal used for the x-axis (xsigName::AbstractString) and the name used for the y-axis (ysigName::AbstractString), return

xsig::Vector{Vector{Matrix{Real}}}: the x-axis signal (the matrix has one column)

ysig::Vector{Vector{Matrix{Real}}}: the y-axis signal.

xsigLegend::AbstractString: the legend of the x-axis signal.

ysigLegend::Vector{AbstractString}: the legend of the y-axis signal as String vector,

ysigType::ModiaResult.SignalType: Type of signal ysig.

xsig[i][j,1] are the x-values at time instant j of segment i

ysig[i][j,:] are the (flattened) array values at time instant j of segment i.

ModiaResult.signalLength — Function

signalLength(signal)

Return the total number of values of signal::Vector{AbstractVector}. If signal[i] is nothing or missing, a length of zero is returned.

ModiaResult.hasSameSegments — Function

hasSameSegments(signal1, signal2)

Return true, if the lengths of the segments in signal1 and in signal2 are the same.

ModiaResult.hasDimensionMismatch — Function

hasDimensionMismatch(signal, timeSignal, timeSignalName)

Print a warning message if signalLength(signal) != signalLength(timeSignal) and return true. Otherwise, return false

ModiaResult.getHeading — Function

getHeading(result, heading)

Return heading if no empty string. Otherwise, return defaultHeading(result).

ModiaResult.prepend! — Function

prepend!(prefix, signalLegend)

Add prefix string in front of every element of the signalLegend string-Vector.